Класс ноды вынесен в отдельный файл
