+++ Source/Applications/Desktop/VodovozViewModels/VodovozViewModels.csproj
@@ -703,6 +703,7 @@
     <Compile Include="TempAdapters\IOrderSelectorFactory.cs" />
     <Compile Include="TempAdapters\SearchHelper.cs" />
     <Compile Include="TempAdapters\SearchParameter.cs" />
+    <Compile Include="Widgets\Cars\CarModelSelection\CarModelSelectionFilterViewModel.CarModelSelectableNode.cs" />
     <Compile Include="Widgets\Cars\CarModelSelection\CarModelSelectionFilterViewModel.cs" />
     <Compile Include="Widgets\Search\CompositeSearchViewModel.cs" />
     <Compile Include="Widgets\UndeliveredOrderViewModel.cs" />
